[fix] 子コンポーネントで定石を取得
